SpreadsheetExpressionReferenceLoaderTesting2 extracted SpreadsheetExpressionReferenceLoaderTesting
